Create De ce folosim ReLu (Rectified Linear Unit)
ReLU 
-> can be used with most types of neural networks.
-> It is recommended by default for both multilayer neural networks (MLP) and convolutional neural networks (CNN).


Why is ReLu good? Rectified Linear Unit (ReLU)
Https://machinelearningmastery.com/rectified-linear-activation-function-for-deep-learning-neural-networks/

+ Sigmoid and hyperbolic tangent activation functions cannot be used in multilayer networks due to the missing gradient problem.
+ The rectified linear activation function exceeds the disappearance gradient, allowing models to learn faster and function better.
+ Rectified linear activation (ReLu) is the default activation when developing Multilayer Perceptron and convolutional neural networks.

Sigmoid activation function (also called the logistic function)
-> The entry into function is transformed into a value between 0.0 and 1.0. Entries that are much larger than 1.0 are
Converted to 1.0, similarly, values ​​much smaller than 0.0 are locked to 0.0. The shape of the function for all
The possible inputs is an S shape from zero to 0.5 to 1.0.

Hyperbolic tangent function (or tanh for short)
-> It is a similar nonlinear activation function that produces values ​​between -1.0 and 1.0.

Sigmoid activation function and hyperbolic tangent function (problem)
-> A general problem with both sigmoid and tanh functions is that they saturate. This means that high values
Are caught at 1.0 and low values ​​are caught at -1 or 0, respectively at tanh and sigmoid. Moreover, the functions are only sensitive to
Changes around the midpoint of their entry, such as 0.5 for sigmoid and 0.0 for tanh.
-> Finally, as hardware capacity has increased through very deep neural networks, using functions
Sigmoid and tanh activation could not be easily instructed.

Resume
-> can be used with most types of neural networks.
-> It is recommended by default for both multilayer neural networks (MLP) and convolutional neural networks (CNN).
-> Try a Smaller Bias Input Value (he bias is the input on the node that has a fixed value.
The bias has the effect of shifting the activation function and it is traditional to set the bias input value to 1.0.
When using ReLU in your network, consider setting the bias to a small value, such as 0.1.
… it can be a good practice to set all elements of [the bias] to a small, positive value, such as 0.1. This makes it very likely 
that the rectified linear units will be initially active for most inputs in the training set and allow the derivatives to pass through.)
-> Use “He Weight Initialization” (Before training a neural network,the weights of the network must be initialized to small random 
values.When using ReLU in your network and initializing weights to small random values centered on zero, then by default half of the 
units in the network will output a zero value.For example, after uniform initialization of the weights, around 50% of hidden units 
continuous output values are real zeros)
-> Scale Input Data (It is good practice to scale input data prior to using a neural network.This may involve standardizing variables 
to have a zero mean and unit variance or normalizing each value to the scale 0-to-1.Without data scaling on many problems, the weights 
of the neural network can grow large, making the network unstable and increasing the generalization error.This good practice of scaling 
inputs applies whether using ReLU for your network or not.)
-> Use Weight Penalty (By design, the output from ReLU is unbounded in the positive domain.This means that in some cases, the output 
can continue to grow in size. As such, it may be a good idea to use a form of weight regularization, such as an L1 or L2 vector norm.
Another problem could arise due to the unbounded behavior of the activations; one may thus want to use a regularizer to prevent 
potential numerical problems. Therefore, we use the L1 penalty on the activation values, which also promotes additional sparsity)
L1 & L2 norm ( https://machinelearningmastery.com/vector-norms-machine-learning/ )
-> Extensions and Alternatives to ReLU ( The ReLU does have some limitations.Key among the limitations of ReLU is the case where large weight updates can mean that the summed input to the activation function is always negative, regardless of the input to the network.This means that a node with this problem will forever output an activation value of 0.0. This is referred to as a “dying ReLU“. )